rename: 클래스명 변경 - #135
- ApiResponse 보다 구체적인 네이밍 적용
